Move CreatePartyActivity to party
