+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PartyMaker">
-        <activity android:name=".CreatePartyActivity"></activity>
+        <activity android:name=".party.CreatePartyActivity"></activity>
         <activity android:name=".user.LoginActivity"></activity>
         <activity android:name=".MainActivity">
             <intent-filter>
