+++ app/src/main/java/com/github/partymakers/partymaker/MainActivity.java
@@ -11,6 +11,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.firebase.ui.auth.AuthUI;
 import com.firebase.ui.auth.IdpResponse;
 import com.github.partymakers.partymaker.databinding.ActivityMainBinding;
+import com.github.partymakers.partymaker.party.CreatePartyActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
