+++ app/src/main/java/com/github/partymakers/partymaker/party/CreatePartyActivity.java
@@ -1,4 +1,4 @@
-package com.github.partymakers.partymaker;
+package com.github.partymakers.partymaker.party;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,7 +7,6 @@ import android.app.TimePickerDialog;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.DatePicker;
-import android.widget.EditText;
 import android.widget.TimePicker;
 
 import com.github.partymakers.partymaker.databinding.ActivityCreatePartyBinding;
@@ -15,7 +14,7 @@ import com.github.partymakers.partymaker.databinding.ActivityCreatePartyBinding;
 import java.util.Calendar;
 
 public class CreatePartyActivity extends AppCompatActivity implements View.OnClickListener {
-    private int mYear, mMonth, mDay, mHour, mMinute;
+    private int year, month, day, hour, minute;
     private ActivityCreatePartyBinding viewBinding;
 
     @Override
@@ -34,25 +33,25 @@ public class CreatePartyActivity extends AppCompatActivity implements View.OnCli
 
         if (v == viewBinding.buttonDatePicker) {
             // Get Current Date
-            final Calendar c = Calendar.getInstance();
-            mYear = c.get(Calendar.YEAR);
-            mMonth = c.get(Calendar.MONTH);
-            mDay = c.get(Calendar.DAY_OF_MONTH);
+            final Calendar calendar = Calendar.getInstance();
+            year = calendar.get(Calendar.YEAR);
+            month = calendar.get(Calendar.MONTH);
+            day = calendar.get(Calendar.DAY_OF_MONTH);
 
             DatePickerDialog datePickerDialog = new DatePickerDialog(this, new DatePickerDialog.OnDateSetListener() {
                 @Override
                 public void onDateSet(DatePicker view, int year, int monthOfYear, int dayOfMonth) {
                     viewBinding.textDatePicked.setText(dayOfMonth + "-" + (monthOfYear + 1) + "-" + year);
                 }
-            }, mYear, mMonth, mDay);
+            }, year, month, day);
             datePickerDialog.show();
         }
 
         if (v == viewBinding.buttonTimePicker) {
             // Get Current Time
-            final Calendar cal = Calendar.getInstance();
-            mHour = cal.get(Calendar.HOUR_OF_DAY);
-            mMinute = cal.get(Calendar.MINUTE);
+            final Calendar calendar = Calendar.getInstance();
+            hour = calendar.get(Calendar.HOUR_OF_DAY);
+            minute = calendar.get(Calendar.MINUTE);
 
             // Launch Time Picker Dialog
             TimePickerDialog timePickerDialog = new TimePickerDialog(this, new TimePickerDialog.OnTimeSetListener() {
@@ -61,7 +60,7 @@ public class CreatePartyActivity extends AppCompatActivity implements View.OnCli
                                       int minute) {
                     viewBinding.textTimePicked.setText(hourOfDay + ":" + minute);
                 }
-            }, mHour, mMinute, false);
+            }, hour, minute, false);
             timePickerDialog.show();
 
             // TODO: date and time validation
